Aggregated data in prep for ordination
These few lines took a few hours because I couldn't figure out why aggregate was giving innaccurate values since it was automaticcaly converting factors to numeric which changes the values. Had to break up the process and slowly work through but found the problem and rememerbed I already learned about that conflict. Fun stuff but we made it