Trying out mapping with different datasets
Please don't judge how messy this is I've been trying a lot of different methods and datasets for mapping and it will be cleaned as I figure out the last tweaks and which datasets we want. I need o figure out why the decimal degree coordinates aren't mapping to the right locations. I can't tell if it has to do with crs or formatting