Optimizing previous script and adding way to find dry periods
Moved parts around to allow this to be as efficient as possible. I still run the freeze loop seperately and merge but I can't think of any way that is more efficient. I also finally got the dry period subsetting to work. I must have tried 50 methods over a full work day but I finally got it and it's painfully simple